requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
scikit-learn==1.7.1
numpy==2.3.1
pandas==2.3.1
Pillow==11.3.0
joblib==1.5.1
python-multipart==0.0.20
